Refactor (APIMF-3491): extracted object collection parsing
